Code review fixes and correctness improvements
- FileDescriptor: replace IDisposable with explicit Close() method
- LinuxString.Format: handle small bufferSize without crashing
- LinuxCancellationToken.None: cache as singleton
- Remove [SuppressGCTransition] from read/write/send/recv/recvfrom
- Fix poll() nfds type to nuint (nfds_t)
- Fix EWOULDBLOCK = TryAgain (EAGAIN alias on Linux)
- Fix O_PATH and O_TMPFILE flag values
- LinuxResult: throw on unsupported type sizes
- NativeObject: thread-safe double-dispose guard

+++ LinuxCore/Interop/Poll.cs
@@ -23,5 +23,5 @@ internal static unsafe partial class Poll
     // int poll(struct pollfd *fds, nfds_t nfds, int timeout);
     [LibraryImport(LinuxLibraries.LibC, EntryPoint = "poll")]
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static partial LinuxResult<int> poll(pollfd* fds, uint nfds, int timeout);
+    public static partial LinuxResult<int> poll(pollfd* fds, nuint nfds, int timeout);
 }
